Moved routes.ts into data folder
Moved routes.ts inside data folder so it look more clear as the logic is inside the hook useRoute and not routes.ts itself.
